added log in and register functionality
