+++ Chapter8_DatabaseInViews/Projects/daniel_gatenadze/templates/base.html
@@ -1,4 +1,4 @@
-{% from "macro.html" import nav_active,nav_active_secondary, nav_collapse_items %}
+{% from "macro.html" import nav_active,nav_active_secondary,nav_activate_logout, nav_collapse_items %}
 <!DOCTYPE html>
 <html lang="en">
 <head>
