+++ Chapter8_DatabaseInViews/Projects/daniel_gatenadze/templates/macro.html
@@ -16,7 +16,6 @@
 {% endmacro %}
 
 
-
 {% macro nav_collapse_items(endpoint, name) %}
     <a class="dropdown-item" href="{{ url_for(endpoint) }}">{{ name }}</a>
 {% endmacro %}
